Add buffer playlist for weekly playlist
Previous logic had weekly playlist starting blank at the beginning of
the week and filling up with songs. This meant that the best time to
listen to the playlist was right before it got wiped.

With the new logic, songs are slowly added to the buffer playlist, and
then at the end of the week all songs are copied from the buffer into
the weekly playlist to make an "All Songs of Last Week" playlist. People
can then listen to this weekly playlist all week long and the playlist
is most interesting right as the buffer is copied over.