Change redirect URI to localhost
In the Spotify Authorization Code Flow in spotipy, the user will be
directed to open a Spotify link and click "Authorize" to authorize the
app. Once authorized, Spotify will redirect the user's browser to the
listed redirect URI and append the authorization code to the URI
as a query string. The user then copies the redirect URI and query
string and pastes it into the terminal to pass the code to spotipy.

Currently, the user is redirected to example.com, which is not under
the control of myself or the user. Conceivably, an attacker with control
of example.com could intercept the Spotify authorization code and gain
access to the user's Spotify account. To prevent the code from being
accidentally leaked, the redirect is changed to localhost, which should
not contact any computers outside of the user's control.