Handle many servers & playlists with single bot
This change is intended to allow a single bot to be added to many
servers. In each server, the bot will watch the desired channel for that
server and add any songs to that server's desired playlists.

Secrets have been split out of config.json so config.json is holding the
info for each guild such as guild id, channel id, and playlist ids. This
should make replacing config.json with a nosql database straightforward.

playlist_update_datetime is bundled with config.json but outside of
guilds at the moment because I haven't yet figured out how to have each
guild run a playlist update at a different time.

.gitignore is updated to exclude secrets*.json files to prevent
accidentally committing secrets.