Store config, secrets, cache in docker vol
Previously, discobot had to run once as standalone python program to
generate valid Spotify OAuth .cache file, which would then be copied
along with config.json and secrets.json into the docker container at
build time. However, this mixes code with configuration by baking the
configuration information into the container during build.

Instead, config.json and secrets.json are now stored in a folder that
is mounted to /discobot/config at container runtime. This also allows
the Spotify OAuth token to be saved between container runs.

The README is updated to reflect creating the docker config folder and
mounting it at runtime. The docker run instruction is also modified to
add the -it option, which allows the user to complete the Spotify
OAuth process in a docker container by pasting the link into the
terminal.

For consistency between running discobot at the command line and
running in a docker container, the config and secrets files are moved
into a config folder in the repository.

In addition, the README is updated to document commit
0187a6c7e4210bb0fafae8f253e447d1d9473bfa, where the secrets and config
were split into separate json files.

The README is also updated to add info on how to add multiple discord
servers to make the process clearer.