Remove command line args, instead use secrets.json
Command line arguments turn out to be a pain when running in docker, as
changing the entrypoint of the docker container to add command line
arguments at runtime doesn't seem to be possible. Instead, the secrets
file will have a standardized location, and the mongodb uri will be
stored in the secrets file along with the username and password.