Add encryption of Spotify tokens in MongoDB
Discobot needs to store the users' Spotify access tokens in order to add
songs to the users' spotify playlists, so the tokens are stored in a
MongoDB store.

However, if an attacker were to gain access to that MongoDB store, they
could use the users' tokens to modify or delete the users playlists. To
prevent an attacker from using the tokens, the tokens are encrypted
at rest in the database.

This moves the weak point to the decryption key. To make the key more
difficult to access, assymetric RSA cryptography is used to separate the
encryption and decryption keys. This way, when the user signs up for
Discobot, the web server handling the signup and user adding process
only needs to have the public key to add entries to the Mongo database
and is unable to decrypt existing entries. If the web server gets
breached, the attackers will then have to move laterally into the
chatbot container to obtain the private key.

RSA encryption and decryption is implemented following this guide:
https://nitratine.net/blog/post/asymmetric-encryption-and-decryption-in-python/

RSA cannot encrypt content larger than the RSA key. The approach
listed to solve this problem is to create a symmetric key, use the
symmetric key to encrypt the data, then encrypt the symmetric key with
the RSA public key and send the encrypted data and encrypted symmetric
key together to the recipient.
https://stackoverflow.com/questions/1199058/how-to-use-rsa-to-encrypt-files-huge-data-in-c-sharp
The cryptography library default Fernet encryption is used for
symmetric encryption.