Add support for multiple users and tokens
In the old code, a single Spotify object authorized for a single
user was used to handle manipulating playlists. However, this means
that the single user must have modification permission on all playlists,
which makes supporting multiple users difficult.

To support multiple users, when discobot begins an action (like
adding a song to a playlist, updating playlists, etc), the relevant
user's information is pulled from MongoDB and used to create a Spotify
object to perform the action.

For example, if a message with a song link is received, discobot
will check the guilds table in MongoDB to determine which user owns the
guild that originated the message. The guild owner's Spotify auth token
will then be used to add the song to the guild owner's playlist.