Organize into chatbot folder & add docker-compose
In anticipation of adding a frontend and backend for discobot, move
the chatbot & related files into its own folder, change the name from
discobot to chatbot, then update the dockerfile to reflect the new
naming and folder structure.

In addition, a docker-compose_debug.yaml file was added to organize &
deploy the various containers. The docker-compose file stores the mongo
data in ./debug_volumes/, so the .gitignore file was updated to ignore
the data stored in ./debug_volumes/.

Finally, ./start.sh was added to make catching the Spotify OAuth link
from the chatbot easier at startup.