Add user doing Spotify OAuth to get code
Spotify doesn't allow changeable query parameters in the redirect URI,
but does allow a unique "state" token to be added to the authorize link
that will be returned along with the authorization code when Spotify
redirects the user to the discobot redirect URI. By connecting the state
token to a profile, the returned authorization code can be matched to a
profile when the user is redirected to the discobot spotify_redirect
url. A spotify_state field has been added to the Profile model to keep
track of the state token.
See https://developer.spotify.com/documentation/general/guides/authorization-guide/#authorization-code-flow

To make sign-up flow easy, we should redirect users with no Spotify
token to set up their Spotify OAuth connection instead of going to the
manage_profile page.

Tests have been modified to reflect the new redirection to the Spotify
OAuth setup page. In addition, tests for the redirection when the user's
profile has a spotify_auth_token are also added.

Tests for the Spotify auth process are also added.