Add homepage login button & post-login redirects
Django allows specifying a redirect URL as a string, but the manage user
page includes the user id. Decided to redirect user to the create_user
page (which is just '/user') and let the create_user page redirect to
the manage_user page including the user id. In other words, the user
logs in, is redirected to '/user', which then redirects them to
'/user/<id>'.

Tests for this redirect behaviour were added. The test classes were also
split up to group like tests together.